refactor: 도메인별 디렉토리로 엔티티 클래스 분리 #29
